Refactored away global state
This commit gets rid of the global state variables in favor of stuffing
them into a data structure. The data structure still lives in the
global scope but it is now theoretically possible for multiple instances
of this plugin to be sharing the same process space. Different contexts
are indexed using the instance handle since it's impossible to attach
user data to an opaque instance handle.