Making decompressor resilient (incomplete)
Trying to make the decompressor operate on arbitrarily large buffers
and make no assumptions about how large the output buffer needs to be.
This is not working yet and is still papered over by a rather large
(1 megabyte) decompression buffer so that the decompressor function
can perform its entire job with one call.