Clear viz bitmap while respecting alpha
It's not enough to clear a bitmap to black by just memset()'ing 0. That
sets alpha values to 0. This doesn't matter on Windows or Linux but causes
problems on Mac OS X. Instead, build a black pixel that sets alpha to 255
and set each pixel in the bitmap to that value.